refactor: NavigationController 통합 및 네비바 관리를 RootTabBarController로 이관
