+++ FoodDiary/Presentation/Sources/Calendar/CalendarViewController.swift
@@ -9,7 +9,6 @@ import UIKit
 import Combine
 import DesignSystem
 import SnapKit
-import DI
 
 public final class CalendarViewController: UIViewController {
     public enum ViewMode {
@@ -20,16 +19,10 @@ public final class CalendarViewController: UIViewController {
     private let weeklyVC: UIViewController
     private let monthlyVC: UIViewController
     private let currentModeSubject = CurrentValueSubject<ViewMode, Never>(.weekly)
-    private let didLogoutSubject = PassthroughSubject<Void, Never>()
     private var currentChild: UIViewController?
-    private var myPageCancellable: AnyCancellable?
 
     public var currentModePublisher: AnyPublisher<ViewMode, Never> {
         currentModeSubject.eraseToAnyPublisher()
-    }
-
-    public var didLogoutPublisher: AnyPublisher<Void, Never> {
-        didLogoutSubject.eraseToAnyPublisher()
     }
 
     public init(weeklyVC: UIViewController, monthlyVC: UIViewController) {
@@ -44,44 +37,7 @@ public final class CalendarViewController: UIViewController {
 
     public override func viewDidLoad() {
         super.viewDidLoad()
-        setupNavigationBar()
         showViewController(for: currentModeSubject.value)
-    }
-
-    public override func viewWillAppear(_ animated: Bool) {
-        super.viewWillAppear(animated)
-        navigationController?.setNavigationBarHidden(false, animated: animated)
-    }
-
-    private func setupNavigationBar() {
-        let mypageButton = UIBarButtonItem(
-            image: DesignSystemAsset.iconMypage.image,
-            style: .plain,
-            target: self,
-            action: #selector(mypageButtonTapped)
-        )
-        navigationItem.rightBarButtonItem = mypageButton
-
-        let logoImageView = UIImageView(image: DesignSystemAsset.iconNavLogo.image)
-        logoImageView.contentMode = .scaleAspectFit
-        navigationItem.leftBarButtonItem = UIBarButtonItem(customView: logoImageView)
-    }
-
-    @objc private func mypageButtonTapped() {
-        guard let myPageVM = try? DIContainer.shared.resolve(MyPageViewModel.self) else {
-            print("MyPageViewModel resolve 실패")
-            return
-        }
-
-        let myPageVC = MyPageViewController(viewModel: myPageVM)
-        myPageVC.hidesBottomBarWhenPushed = true
-
-        myPageCancellable = myPageVC.didLogoutPublisher
-            .sink { [weak self] in
-                self?.didLogoutSubject.send()
-            }
-
-        navigationController?.pushViewController(myPageVC, animated: true)
     }
 
     public func toggleViewMode() {
