+++ FoodDiary/Presentation/Sources/TabBar/RootTabBarController.swift
@@ -7,32 +7,41 @@
 
 import UIKit
 import DesignSystem
+import DI
 import Combine
 
 public final class RootTabBarController: UITabBarController {
     let calendarVC: CalendarViewController
-    let insightVC: UIViewController
+    let insightVC: InsightViewController
+    private let didLogoutSubject = PassthroughSubject<Void, Never>()
+    private var myPageCancellable: AnyCancellable?
     private var cancellables = Set<AnyCancellable>()
 
     public var didLogoutPublisher: AnyPublisher<Void, Never> {
-        calendarVC.didLogoutPublisher
+        didLogoutSubject.eraseToAnyPublisher()
     }
 
-    public init(weeklyVC: UIViewController, monthlyVC: UIViewController, insightVC: UIViewController) {
+    public init(weeklyVC: UIViewController, monthlyVC: UIViewController, insightVC: InsightViewController) {
         self.calendarVC = CalendarViewController(weeklyVC: weeklyVC, monthlyVC: monthlyVC)
         self.insightVC = insightVC
         super.init(nibName: nil, bundle: nil)
     }
-    
+
     required init?(coder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
     }
-    
+
     public override func viewDidLoad() {
         super.viewDidLoad()
         setup()
+        setupNavigationBar()
     }
-    
+
+    public override func viewWillAppear(_ animated: Bool) {
+        super.viewWillAppear(animated)
+        navigationController?.setNavigationBarHidden(false, animated: animated)
+    }
+
     private func setup() {
         self.delegate = self
         tabBar.tintColor = DesignSystemAsset.primary.color
@@ -44,30 +53,44 @@ public final class RootTabBarController: UITabBarController {
             }
             .store(in: &cancellables)
 
-        let calendarNav = UINavigationController(rootViewController: calendarVC)
-        calendarNav.tabBarItem = UITabBarItem(title: "홈", image: DesignSystemAsset.iconHome.image, tag: 0)
-
-        let navAppearance = UINavigationBarAppearance()
-        navAppearance.configureWithOpaqueBackground()
-        navAppearance.backgroundColor = .sdBase
-        navAppearance.titleTextAttributes = [.foregroundColor: UIColor.white]
-        navAppearance.shadowColor = .clear
-        
-        calendarNav.navigationBar.standardAppearance = navAppearance
-        calendarNav.navigationBar.scrollEdgeAppearance = navAppearance
-        calendarNav.navigationBar.tintColor = .white
-
-        let insightNav = UINavigationController(rootViewController: insightVC)
-        insightNav.tabBarItem = UITabBarItem(title: "인사이트", image: DesignSystemAsset.iconInsight.image, tag: 1)
-        insightNav.navigationBar.standardAppearance = navAppearance
-        insightNav.navigationBar.scrollEdgeAppearance = navAppearance
-        insightNav.navigationBar.tintColor = .white
+        calendarVC.tabBarItem = UITabBarItem(title: "홈", image: DesignSystemAsset.iconHome.image, tag: 0)
+        insightVC.tabBarItem = UITabBarItem(title: "인사이트", image: DesignSystemAsset.iconInsight.image, tag: 1)
 
         let toggleVC = UIViewController()
         toggleVC.tabBarItem = UITabBarItem(tabBarSystemItem: .search, tag: 2)
         toggleVC.tabBarItem.image = DesignSystemAsset.iconWeekly.image
 
-        viewControllers = [calendarNav, insightNav, toggleVC]
+        viewControllers = [calendarVC, insightVC, toggleVC]
+    }
+
+    private func setupNavigationBar() {
+        let mypageButton = UIBarButtonItem(
+            image: DesignSystemAsset.iconMypage.image,
+            style: .plain,
+            target: self,
+            action: #selector(mypageButtonTapped)
+        )
+        navigationItem.rightBarButtonItem = mypageButton
+
+        let logoImageView = UIImageView(image: DesignSystemAsset.iconNavLogo.image)
+        logoImageView.contentMode = .scaleAspectFit
+        navigationItem.leftBarButtonItem = UIBarButtonItem(customView: logoImageView)
+    }
+
+    @objc private func mypageButtonTapped() {
+        guard let myPageVM = try? DIContainer.shared.resolve(MyPageViewModel.self) else {
+            print("MyPageViewModel resolve 실패")
+            return
+        }
+
+        let myPageVC = MyPageViewController(viewModel: myPageVM)
+
+        myPageCancellable = myPageVC.didLogoutPublisher
+            .sink { [weak self] in
+                self?.didLogoutSubject.send()
+            }
+
+        navigationController?.pushViewController(myPageVC, animated: true)
     }
     
     private func toggleViewMode() {
